feat: add ansible collection downloads count badge
